requirements.txt
feedparser==6.0.11
requests==2.31.0
anthropic==0.25.0
tweepy==4.14.0
python-dotenv==1.0.1
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.1.0
